requirements.txt
fastapi
uvicorn
python-multipart
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch
numpy
pillow
streamlit
requests
opencv-python-headless